requirements.txt
# HTTP Requests
requests==2.28.2

# Data Manipulation and Analysis
pandas==2.0.2
numpy==1.25.1

# Working with APIs
requests-oauthlib==1.3.1

# File and Data Compression
pyyaml==6.0

# Environment Management
virtualenv==20.21.0
python-dotenv==1.0.0

# Logging
loguru==0.7.0

# Smartsheet SDK
smartsheet-python-sdk==2.166.1